text-semantic-search/requirements.txt
google-cloud-dataflow==2.5
google-cloud-storage==1.13.2
tensorflow==1.12
tensorflow-hub==0.2
tensorflow-transform==0.11
annoy==1.15.0
google-api-python-client